requirements.txt
streamlit
pandas
matplotlib
requests
python-dotenv
keyring
openpyxl
pvlib
